# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Atlas Billing API
  version: 0.2.0
  description: Subscription billing — products, prices, coupons, subscriptions, invoices, and credit notes
  contact:
    name: AtlasPayments Team
    url: https://github.com/shivemind/atlaspayments
servers:
  - url: http://localhost:3002
    description: Local billing service. Override via Postman {{baseUrl}} environment variable.
...
paths:
  /api/v1/products:
    get:
      summary: List products
      description: Returns a paginated list of products for the authenticated merchant.
      x-platform:
        kind: resource
        entity: Product
        affectsLedger: false
      operationId: listProducts
      tags:
        - Products
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - in: query
          name: status
          description: Filter by product status (active or archived).
          schema:
            type: string
            enum:
              - active
              - archived
      responses:
        "200":
          description: Paginated list of merchant products.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create product
      description: Creates a new product in the merchant's catalog.
      x-platform:
        kind: resource
        entity: Product
        emitsEvents:
          - product.created
        affectsLedger: false
      operationId: createProduct
      tags:
        - Products
      security:
        - MerchantKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 200
                  description: Product display name.
                description:
                  type: string
                  maxLength: 2000
                metadata:
                  type: object
                  additionalProperties: true
            examples:
              basic:
                value:
                  name: Premium Plan
                  description: Access to all premium features
                  metadata:
                    tier: premium
      responses:
        "201":
          description: Product created successfully.
          content:
            application/json:
              schema:
                type: object
                required:
                  - product
                properties:
                  product:
                    type: object
              examples:
                created:
                  value:
                    product:
                      id: prod_abc123
                      merchant_id: cmf1merchant123
                      name: Premium Plan
                      description: Access to all premium features
                      status: active
                      metadata:
                        tier: premium
                      created_at: 2026-03-01T10:00:00.000Z
                      updated_at: 2026-03-01T10:00:00.000Z
        "400":
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/products/{id}:
    get:
      summary: Get product by ID
      description: Returns a single product scoped to the authenticated merchant.
      x-platform:
        kind: resource
        entity: Product
        affectsLedger: false
      operationId: getProductById
      tags:
        - Products
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Product found.
          content:
            application/json:
              schema:
                type: object
                required:
                  - product
                properties:
                  product:
                    type: object
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Product not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      summary: Update product
      description: Updates a product's name, description, or metadata. Cannot update archived products.
      x-platform:
        kind: transition
        entity: Product
        emitsEvents:
          - product.updated
        affectsLedger: false
        stateChanges:
          - metadata_updated
      operationId: updateProduct
      tags:
        - Products
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 200
                description:
                  type: string
                  maxLength: 2000
                  nullable: true
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Product updated.
          content:
            application/json:
              schema:
                type: object
                required:
                  - product
                properties:
                  product:
                    type: object
        "400":
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Product not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Product is archived.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Archive product
      description: Archives a product (soft delete). Archived products cannot be updated.
      x-platform:
        kind: transition
        entity: Product
        stateChanges:
          - active -> archived
        emitsEvents:
          - product.archived
        affectsLedger: false
      operationId: archiveProduct
      tags:
        - Products
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Product archived.
          content:
            application/json:
              schema:
                type: object
                required:
                  - product
                properties:
                  product:
                    type: object
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Product not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Product is already archived.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/promotion_codes/{id}:
    get:
      summary: Get promotion code by ID
      description: Returns a single promotion code scoped to the authenticated merchant.
      x-platform:
        kind: resource
        entity: PromotionCode
        affectsLedger: false
      operationId: getPromotionCodeById
      tags:
        - PromoCodes
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Promotion code found.
          content:
            application/json:
              schema:
                type: object
                required:
                  - promotion_code
                properties:
                  promotion_code:
                    type: object
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Promotion code not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      summary: Update promotion code
      description: Updates a promotion code's active status or metadata.
      x-platform:
        kind: transition
        entity: PromotionCode
        emitsEvents:
          - promotion_code.updated
        affectsLedger: false
        stateChanges:
          - active -> inactive
          - metadata_updated
      operationId: updatePromotionCode
      tags:
        - PromoCodes
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                is_active:
                  type: boolean
                metadata:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Promotion code updated.
          content:
            application/json:
              schema:
                type: object
                required:
                  - promotion_code
                properties:
                  promotion_code:
                    type: object
        "400":
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Promotion code not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    MerchantKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for merchant-scoped operations. Pass as `Authorization: Bearer {{merchantApiKey}}`.
    PlatformAdminKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for platform admin operations. Pass as `Authorization: Bearer {{platformAdminKey}}`.
  schemas:
    ErrorResponse:
      description: Standard error response envelope returned for 4xx/5xx responses.
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code.
            message:
              type: string
              description: Human-readable error description.
    PaginationMeta:
      description: Pagination metadata included in list responses.
      type: object
      required:
        - page
        - pageSize
        - total
        - hasMore
      properties:
        page:
          type: integer
          description: Current page number (1-indexed).
        pageSize:
          type: integer
          description: Number of results per page.
        total:
          type: integer
          description: Total number of matching records.
        hasMore:
          type: boolean
          description: Whether additional pages exist.
...
  parameters:
...
    PageParam:
      in: query
      name: page
      description: Page number (1-indexed).
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      in: query
      name: pageSize
      description: Number of results per page (max 100).
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    PathId:
      in: path
      name: id
      required: true
      description: Resource identifier.
      schema:
        type: string
...